Fixed string.
2003-01-20 Damien Sandras  <dsandras@seconix.com>

        * gnomemeeting.schemas.in.in: Fixed string.

        * src/common.h: Removed unused widgets in the GmDruidWindow.

        * src/config.cpp: Added notifier. When the hostname changes,
        or the registering method, and when they become "non-mt", then
        disable "microtelco".

        * src/druid.cpp: Changed strings.

        * src/endpoint.cpp: Code cleanup.

        * src/gatekeeper.cpp: Set the registering method to 0 when the
        GK registering fails.

        * src/urlhandler.cpp: Add prefix when MicroTelco is used and
        G.723.1 is not found.

+++ src/endpoint.cpp
@@ -1252,7 +1252,7 @@ GMH323EndPoint::OnConnectionEstablished (H323Connection & connection,
     video_codec->SetTxMaxQuality (vq);
     video_codec->SetBackgroundFill (bf);   
     video_codec->SetMaxBitRate (bitrate * 8 * 1024);
-    video_codec->SetTargetFrameTimeMs (frame_time);
+    video_codec->SetTargetFrameTimeMs ((unsigned int) frame_time);
     video_codec->SetVideoMode (H323VideoCodec::DynamicVideoQuality | 
 			       H323VideoCodec::AdaptivePacketDelay |
 			       video_codec->GetVideoMode());
